Add vanilla cream
The original instructions did not mention vanilla cream. The ingredient
list did, but the vanilla cream ingredients were left out of the printed
copy that was transcribed. I've tried both with and without vanilla
cream and consider it worth mentioning, not to mention more in the
spirit of kanelstænger, but in no way critical.